Move story points hover card
